# api_gateway/api/docs/swagger.yaml
definitions:
...
  forum.CategoryRes:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  forum.PopularTag:
    properties:
      count:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
...
  forum.PostGetAllRes:
    properties:
      post:
        items:
          $ref: '#/definitions/forum.PostRes'
        type: array
    type: object
  forum.PostRes:
    properties:
      body:
        type: string
      category:
        $ref: '#/definitions/forum.CategoryRes'
      id:
        type: string
      title:
        type: string
      user:
        $ref: '#/definitions/forum.UserResp'
    type: object
  forum.PostTags:
    properties:
      id:
        type: string
      post_id:
        type: string
      tag_id:
        type: string
    type: object
  forum.PostTagsCreateReq:
    properties:
      post_id:
        type: string
      tag_id:
        type: string
    type: object
  forum.PostTagsGetAllRes:
    properties:
      posttag:
        items:
          $ref: '#/definitions/forum.PostTagsRes'
        type: array
    type: object
  forum.PostTagsRes:
    properties:
      id:
        type: string
      post:
        $ref: '#/definitions/forum.PostRes'
      tag:
        $ref: '#/definitions/forum.TagRes'
    type: object
  forum.TagCreateReq:
    properties:
      name:
        type: string
    type: object
  forum.TagGetAllRes:
    properties:
      tag:
        items:
          $ref: '#/definitions/forum.TagRes'
        type: array
    type: object
  forum.TagRes:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  forum.UserResp:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /post_tag:
    post:
      consumes:
      - application/json
      description: Create a new post_tag
      parameters:
      - description: Post data
        in: body
        name: post_tag
        required: true
        schema:
          $ref: '#/definitions/forum.PostTagsCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostTags'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create post_tag
      tags:
      - post_tag
  /post_tag/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post_tag by ID
      parameters:
      - description: PostTag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PostTag deleted
          schema:
            type: string
        "400":
          description: Invalid post_tag ID
          schema:
            type: string
        "404":
          description: PostTag not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete post_tag
      tags:
      - post_tag
    get:
      consumes:
      - application/json
      description: Get a post_tag by ID
      parameters:
      - description: PostTag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostTagsRes'
        "400":
          description: Invalid post_tag ID
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get post_tag
      tags:
      - post_tag
    put:
      consumes:
      - application/json
      description: Update an existing post_tag
      parameters:
      - description: PostTag ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated post_tag data
        in: body
        name: post_tag
        required: true
        schema:
          $ref: '#/definitions/forum.PostTagsCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostTags'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: PostTag not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update post_tag
      tags:
      - post_tag
  /post_tags:
    get:
      consumes:
      - application/json
      description: Get all post_tags
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostTagsGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all post_tags
      tags:
      - post_tag
...
  /tag:
    post:
      consumes:
      - application/json
      description: Create a new tag
      parameters:
      - description: Tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/forum.TagCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.TagRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create tag
      tags:
      - tag
  /tag/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tag by ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag deleted
          schema:
            type: string
        "400":
          description: Invalid tag ID
          schema:
            type: string
        "404":
          description: Tag not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete tag
      tags:
      - tag
    get:
      consumes:
      - application/json
      description: Get a tag by ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.TagRes'
        "400":
          description: Invalid tag ID
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get tag
      tags:
      - tag
    put:
      consumes:
      - application/json
      description: Update an existing tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/forum.TagCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.TagRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Comment not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update tag
      tags:
      - tag
  /tags:
    get:
      consumes:
      - application/json
      description: Get all tags
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.TagGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all tags
      tags:
      - tag
  /tags/{tag_id}/posts:
    get:
      consumes:
      - application/json
      description: Get all posts by tag
      parameters:
      - description: TagId
        in: query
        name: tag_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PostGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all posts by tag
      tags:
      - tag
  /tags/popular:
    get:
      consumes:
      - application/json
      description: Get popular tag
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forum.PopularTag'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get popular tag
      tags:
      - tag
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"